Adding common header for Poisson eqn BCs so they are shared between the various Poisson solvers. No enum or structs or functions should be redefined!
