+++ zero/gkyl_fem_poisson_perp.h
@@ -1,35 +1,22 @@
 #pragma once
 
 #include <math.h>
-#include <gkyl_array.h>
-#include <gkyl_basis.h>
-#include <gkyl_range.h>
+
 #include <gkyl_alloc.h>
-#include <gkyl_rect_grid.h>
-#include <gkyl_rect_decomp.h>
+#include <gkyl_array.h>
 #include <gkyl_array_ops.h>
+#include <gkyl_basis.h>
+#include <gkyl_dg_bin_ops.h>
+#include <gkyl_fem_poisson_bctype.h>
 #include <gkyl_mat.h>
 #include <gkyl_mat_triples.h>
-#include <gkyl_dg_bin_ops.h>
+#include <gkyl_range.h>
+#include <gkyl_rect_decomp.h>
+#include <gkyl_rect_grid.h>
 
 // Object type
 typedef struct gkyl_fem_poisson_perp gkyl_fem_poisson_perp;
 
-// Boundary condition types.
-enum gkyl_poisson_bc_type {
-  GKYL_POISSON_PERIODIC=0,
-  GKYL_POISSON_DIRICHLET,  // sets the value. 
-  GKYL_POISSON_NEUMANN,  // sets the slope normal to the boundary.
-};
-
-// Boundary condition values. Dirichlet and Neumann use only one value,
-// Robin uses 3, and periodic ignores the value.
-struct gkyl_poisson_bc_value { double v[3]; };
-
-struct gkyl_poisson_bc {
-  enum gkyl_poisson_bc_type lo_type[GKYL_MAX_CDIM], up_type[GKYL_MAX_CDIM];
-  struct gkyl_poisson_bc_value lo_value[GKYL_MAX_CDIM], up_value[GKYL_MAX_CDIM];
-};
 
 /**
  * Create new updater to solve the Helmholtz problem
@@ -77,31 +64,3 @@ void gkyl_fem_poisson_perp_solve_cu(gkyl_fem_poisson_perp* up, struct gkyl_array
  * @param up Updater to delete.
  */
 void gkyl_fem_poisson_perp_release(struct gkyl_fem_poisson_perp *up);
-
-GKYL_CU_DH
-static inline int idx_to_inup_ker(const int dim, const int *num_cells, const int *idx) {
-  // Return the index of the kernel (in the array of kernels) needed given the grid index.
-  // This function is for kernels that differentiate between upper cells and
-  // elsewhere.
-  int iout = 0;
-  for (int d=0; d<dim; d++) {
-    if (idx[d] == num_cells[d]) iout += (int)(pow(2,d)+0.5);
-  }
-  return iout;
-}
-
-GKYL_CU_DH
-static inline int idx_to_inloup_ker(const int dim, const int *num_cells, const int *idx) {
-  // Return the index of the kernel (in the array of kernels) needed given the grid index.
-  // This function is for kernels that differentiate between lower, interior
-  // and upper cells.
-  int iout = 0;
-  for (int d=0; d<dim; d++) {
-    if (idx[d] == 1) {
-      iout = 2*iout+(int)(pow(3,d)+0.5);
-    } else if (idx[d] == num_cells[d]) {
-      iout = 2*iout+(int)(pow(3,d)+0.5)+1;
-    }
-  }
-  return iout;
-}
